refactor: workflow API 구조 정리
