+++ src/entities/workflow/api/create-workflow.api.ts
@@ -1,4 +1,4 @@
-import { request } from "../core";
+import { request } from "@/shared/api/core";
 
 import type { CreateWorkflowRequest, WorkflowResponse } from "./types";
