+++ src/entities/workflow/api/delete-workflow-node.api.ts
@@ -1,4 +1,4 @@
-import { request } from "../core";
+import { request } from "@/shared/api/core";
 
 import type { WorkflowResponse } from "./types";
