+++ src/entities/workflow/api/get-workflow.api.ts
@@ -1,4 +1,4 @@
-import { request } from "../core";
+import { request } from "@/shared/api/core";
 
 import type { WorkflowResponse } from "./types";
