+++ src/entities/workflow/api/select-workflow-choice.api.ts
@@ -1,4 +1,4 @@
-import { request } from "../core";
+import { request } from "@/shared/api/core";
 
 import type { NodeChoiceSelectRequest, NodeSelectionResult } from "./types";
