+++ src/entities/workflow/api/update-workflow-node.api.ts
@@ -1,4 +1,4 @@
-import { request } from "../core";
+import { request } from "@/shared/api/core";
 
 import type { NodeUpdateRequest, WorkflowResponse } from "./types";
