+++ src/entities/workflow/index.ts
@@ -1,1 +1,2 @@
+export * from "./api";
 export * from "./model";
