+++ src/shared/api/index.ts
@@ -1,3 +1,2 @@
 export * from "./client";
-export * from "./workflow";
 export * from "./query-policy";
