+++ src/shared/model/workflow/useWorkflowChoicesQuery.ts
@@ -1,6 +1,6 @@
 import { useQuery } from "@tanstack/react-query";
 
-import { workflowApi } from "../../api";
+import { workflowApi } from "@/entities/workflow";
 import { workflowKeys } from "../../constants";
 import {
   type QueryPolicyOptions,
